Moves FormFieldTable
